requirements.txt
deep-translator==1.9.2
Pyrogram==2.0.97
TgCrypto==1.2.5
SQLAlchemy==2.0.0
python-dotenv==0.21.1
langdetect==1.0.9
# dev
black==22.12.0